+++ src/main/java/com/malikov/shopsystem/web/controller/OrderAjaxController.java
@@ -1,6 +1,5 @@
 package com.malikov.shopsystem.web.controller;
 
-import com.malikov.shopsystem.model.OrderItem;
 import com.malikov.shopsystem.model.OrderStatus;
 import com.malikov.shopsystem.model.PaymentType;
 import com.malikov.shopsystem.service.CustomerService;
@@ -15,6 +14,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
+
+import java.math.BigDecimal;
 
 @RestController
 @RequestMapping(value = "/ajax/profile/orders")
@@ -119,7 +120,7 @@ public class OrderAjaxController {
     }
 
     @PutMapping(value = "/{orderId}/total-sum")
-    public void updateTotalSum(@PathVariable("orderId") Long orderId, @RequestParam("totalSum") Integer totalSum) {
+    public void updateTotalSum(@PathVariable("orderId") Long orderId, @RequestParam("totalSum") BigDecimal totalSum) {
         orderService.updateTotalSum(orderId, totalSum);
     }
